WIP
Now gets into the loading of the cpu description.  I had to nuke some
of the exception handling code in guile2.scm, which probably needs to
be revisited.

I'm now stuck on two uses of `procedure->memoizing-macro' in cos.scm
that I'm going to have to figure out and remove.